chore: restructure code - dashboard views - courses
